moved more integ tests to xpack module
